python/requirements.txt
confini>=0.3.6rc3,<0.5.0
chainlib-eth>=0.0.10a6,<0.1.0
funga-eth==0.5.1a2
